Improved new static encryption system
The demo algorithm security level is pretty much unknown but the proof-of-concept is a good step forward. It is now possible to do multiple encryption passes.

Encryption is constexpr while Decryption is not (to prevent compile-time encrypt/decrypt sequence from being optimized out). The key should ideally be exported independently to the encrypted data.